ERP-MidCap_BugSolve :: Master Page Toast Message show

+++ MidCapERP.Admin/wwwroot/js/Company.js
@@ -24,7 +24,7 @@ $(function () {
                 "mData": null, "bSortable": false,
                 "mRender": function (o) {
                     return '<div class="c-action-btn-group justify-content-end"><a data-ajax-complete="CompanyModel.onComplete" data-ajax="true" class="btn btn-icon btn-outline-primary" data-ajax-mode="replace" data-ajax-update="#divUpdateCompany" href="/Company/Update/' + o.lookupValueId + '"><i class="bx bxs-pencil"></i></a>' +
-                        '<a data-ajax-complete="CompanyModel.onDelete" data-ajax="true" class="btn btn-icon btn-outline-danger" data-ajax-mode="replace" href="/Company/Delete/' + o.lookupValueId + '"><i class="bx bxs-trash"></i></a></div>';
+                        '<a data-ajax-complete="CompanyModel.onDelete" data-ajax="true" data-ajax-confirm="Are you sure you want to delete?" class="btn btn-icon btn-outline-danger" data-ajax-mode="replace" href="/Company/Delete/' + o.lookupValueId + '"><i class="bx bxs-trash"></i></a></div>';
                 }
             }
         ]
@@ -45,15 +45,17 @@ CompanyModel.onComplete = function () {
 }
 
 CompanyModel.onDelete = function () {
-    tblCompany.ajax.reload();
+    tblCompany.ajax.reload(null, false);
+    toastr.error('Data deleted successfully.');
 }
 
 CompanyModel.onSuccess = function (xhr) {
-    tblCompany.ajax.reload();
+    tblCompany.ajax.reload(null, false);
     $("#divCompanyModal").modal('hide');
+    toastr.success('Information saved successfully.');
 };
 
 CompanyModel.onFailed = function (xhr) {
-    tblCompany.ajax.reload();
+    tblCompany.ajax.reload(null, false);
     $("#divCompanyModal").modal('hide');
 };
